fix(YouTube): Merge `Hide filter bar` patch into `Hide layout components`
