+++ src/main/kotlin/app/revanced/patches/youtube/layout/hide/general/fingerprints/RelatedChipCloudFingerprint.kt
@@ -1,7 +1,7 @@
-package app.revanced.patches.youtube.layout.hide.filterbar.fingerprints
+package app.revanced.patches.youtube.layout.hide.general.fingerprints
 
 import app.revanced.patcher.extensions.or
-import app.revanced.patches.youtube.layout.hide.filterbar.HideFilterBarResourcePatch
+import app.revanced.patches.youtube.layout.hide.general.HideLayoutComponentsResourcePatch
 import app.revanced.util.patch.LiteralValueFingerprint
 import com.android.tools.smali.dexlib2.AccessFlags
 import com.android.tools.smali.dexlib2.Opcode
@@ -14,5 +14,5 @@ internal object RelatedChipCloudFingerprint : LiteralValueFingerprint(
         Opcode.INVOKE_VIRTUAL,
         Opcode.MOVE_RESULT_OBJECT
     ),
-    literalSupplier = { HideFilterBarResourcePatch.relatedChipCloudMarginId }
+    literalSupplier = { HideLayoutComponentsResourcePatch.relatedChipCloudMarginId }
 )
